+++ src/GRA.Data/Repository/TriggerRepository.cs
@@ -336,13 +336,13 @@ namespace GRA.Data.Repository
             return requirements.OrderBy(_ => _.Name).ToList();
         }
 
-        public async Task<int> CountRequirementsAsync(TriggerFilter filter)
+        public async Task<int> CountRequirementsAsync(BaseFilter filter)
         {
             return await ApplyRequirementsFilters(filter)
                 .CountAsync();
         }
 
-        public async Task<ICollection<TriggerRequirement>> PageRequirementsAsync(TriggerFilter filter)
+        public async Task<ICollection<TriggerRequirement>> PageRequirementsAsync(BaseFilter filter)
         {
             return await ApplyRequirementsFilters(filter)
                 .OrderBy(_ => _.Name)
@@ -350,7 +350,7 @@ namespace GRA.Data.Repository
                 .ToListAsync();
         }
 
-        private IQueryable<TriggerRequirement> ApplyRequirementsFilters(TriggerFilter filter)
+        private IQueryable<TriggerRequirement> ApplyRequirementsFilters(BaseFilter filter)
         {
             // Badge and Trigger lists
             var requirements = (from challenges in _context.Challenges